move printer under Printer namespace
